pyproject.toml
[project]
dependencies = [
  "aiohttp==3.9.5",
  "apscheduler>=3.11.0",
  "ccxt>=4.4.59",
  "discord>=2.3.2",
  "load-dotenv>=0.1.0",
  "nest-asyncio>=1.6.0",
  "pandas>=2.2.3",
  "pre-commit>=4.1.0",
  "ruff>=0.9.6",
]
description = "Add your description here"
name = "quant-hub"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"
